Gid rid of bash specific escapes for black_args
Tried some experiments in the same Docker container / bash script that black runs in, and it seems to add single quotes around the last argument, which was telling to black to match files that started with double quotes, or had back slashes in their names. Not sure why it does that, but it looks as if the string is copied directly into bash, no need for escapes.