Fixes incorrect error for DOCX 0th variables
If you have a possible person variable, the Weaver tries to map the string to a python variable.
However, it doesn't isn't passing around the document type where it should, so it considers some DOCX
variables as PDF variables, and will error if you include a valid 0-indexed DOCX person varibable. This
adds the correct document type to the proper functions.