Don't tell DOCX users they can't use 0 idx in PDF
Happens because we try to detect if they are using PDF field names, but the
check for a 0 index (i.e. "users0") happens in that same function. Fixed
by just skipping ParsingExceptions: they are by definition all PDF parsing
errors, which we don't really need to handle, as we already find DOCX parsing
issues.