If looking for custom people, don't use _name
Was the reason that we turned off custom people for LITCon. Before we turn it
back on, we should stop the cause of false positives.

Decided to hardcode the ending to avoid "_name" in the method for now; I don't
think we need to make it general now, as it's kinda hacky, and still needs a
better long term syntax to handle when people want things to be objects or
people, like #96.