Be able to ignore fields with bad names
Can across some PDFs that had fields had bad Adobe names (like \`25, etc.), but
Gavel wouldn't show those fields, so authors with those PDFs aren't able to
progress.

We have the ability to just not fill in certain fields in the docassemble, and
this patch allows authors to continue with the Weaving process if they don't
care about the shown fields.